Refactored files and added comments
