+++ app/src/main/res/layout/activity_quiz.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent" android:layout_height="match_parent">
-
 <TextView
     android:id="@+id/EconsIG"
     android:layout_width="wrap_content"
