+++ app/src/main/res/values/strings.xml
@@ -1,17 +1,21 @@
 <resources>
     <string name="app_name">STAPP</string>
+    <!-- sign in string -->
     <string name="Students">Students</string>
     <string name="Teachers">Teachers</string>
+
     <string name="Logout">Logout</string>
     <string name="TODO">TODO</string>
     <string name="title_home">Home</string>
+    <!-- main menu (Drawer string -->
     <string name="openDrawer" />
     <string name="closeDrawer" />
+
     <string name="loading">Loading</string>
     <string name="signed_out">Successfully signed out</string>
     <string name="delete_todo">Delete TODO</string>
 
-    <!--Sign In-->
+    <!--Sign In process-->
     <string name="signed_in_fmt">Signed in as: %s</string>
 
     <!-- unknown value -->
